Promote elasticsearch to native support
- Move elasticsearch extension from extensions-jvm to extensions
- Move integration tests from integration-tests-jvm to integration-tests
- using quarkus-elasticsearch-java-client to support camel-quarkus-elasticsearch native mode
- All tests pass in both JVM and native modes (9/9)

Fixes #4155
