+++ docs/modules/ROOT/examples/components/elasticsearch.yml
@@ -2,11 +2,11 @@
 # This file was generated by camel-quarkus-maven-plugin:update-extension-doc-page
 cqArtifactId: camel-quarkus-elasticsearch
 cqArtifactIdBase: elasticsearch
-cqNativeSupported: false
-cqStatus: Preview
+cqNativeSupported: true
+cqStatus: Stable
 cqDeprecated: false
 cqJvmSince: 3.2.0
-cqNativeSince: n/a
+cqNativeSince: 3.36.0
 cqCamelPartName: elasticsearch
 cqCamelPartTitle: Elasticsearch
 cqCamelPartDescription: Send requests to ElasticSearch via Java Client API.
