+++ docs/modules/ROOT/pages/reference/extensions/elasticsearch.adoc
@@ -4,17 +4,17 @@
 = Elasticsearch
 :linkattrs:
 :cq-artifact-id: camel-quarkus-elasticsearch
-:cq-native-supported: false
-:cq-status: Preview
-:cq-status-deprecation: Preview
+:cq-native-supported: true
+:cq-status: Stable
+:cq-status-deprecation: Stable
 :cq-description: Send requests to ElasticSearch via Java Client API.
 :cq-deprecated: false
 :cq-jvm-since: 3.2.0
-:cq-native-since: n/a
+:cq-native-since: 3.36.0
 
 ifeval::[{doc-show-badges} == true]
 [.badges]
-[.badge-key]##JVM since##[.badge-supported]##3.2.0## [.badge-key]##Native##[.badge-unsupported]##unsupported##
+[.badge-key]##JVM since##[.badge-supported]##3.2.0## [.badge-key]##Native since##[.badge-supported]##3.36.0##
 endif::[]
 
 Send requests to ElasticSearch via Java Client API.
@@ -28,6 +28,10 @@ Please refer to the above link for usage and configuration details.
 
 [id="extensions-elasticsearch-maven-coordinates"]
 == Maven coordinates
+
+https://{link-quarkus-code-generator}/?extension-search=camel-quarkus-elasticsearch[Create a new project with this extension on {link-quarkus-code-generator}, window="_blank"]
+
+Or add the coordinates to your existing project:
 
 [source,xml]
 ----
