+++ extensions-jvm/pom.xml
@@ -57,7 +57,6 @@
         <module>drill</module>
         <module>dsl-modeline</module>
         <module>ehcache</module>
-        <module>elasticsearch</module>
         <module>fastjson</module>
         <module>flink</module>
         <module>google-functions</module>
