+++ integration-tests-jvm/pom.xml
@@ -56,7 +56,6 @@
         <module>drill</module>
         <module>dsl-modeline</module>
         <module>ehcache</module>
-        <module>elasticsearch</module>
         <module>fastjson</module>
         <module>flink</module>
         <module>google-functions</module>
