+++ integration-tests/elasticsearch/src/main/java/org/apache/camel/quarkus/component/elasticsearch/it/ElasticsearchRoutes.java
